requirements.txt
Django==1.7.1
MySQL-python>=1.2.5
PyYAML>=3.12
IPy
djangorestframework==3.3.1
markdown==2.6.7
django-filter==0.11.0
